fix: circular type dependencies
